Code refactoring for server package
commit

Code refactoring for server package
