+++ packages/server/src/verifier/verifyPrivacy.ts
@@ -1,0 +1,17 @@
+import { AuthType, hashProperty, Property, Signature } from '@0auth/message';
+
+import { getMerkleRoot, verifyByKeyType } from '../util';
+import { PublicKey } from '../type';
+
+export function verifyPrivacy(
+  properties: Property[],
+  sign: Signature,
+  publicKey: PublicKey,
+): boolean {
+  if (sign.authType !== AuthType.Privacy || sign.keyType !== publicKey.type) {
+    return false;
+  }
+  const hashes = properties.map((property) => hashProperty(property));
+  const merkleRoot = getMerkleRoot(hashes);
+  return verifyByKeyType(merkleRoot, sign.value, publicKey.key, publicKey.type);
+}
