+++ packages/server/tsconfig.json
@@ -5,7 +5,5 @@
   },
   "extends": "../../tsconfig.json",
   "include": ["src/**/*"],
-  "references": [
-    { "path": "../message" }
-  ]
+  "references": [{ "path": "../message" }]
 }
